Random speed improvements (#234)
* Don't check bounds when indexing with enums

The type guarantees they're in bounds

* Use transmutes to convert indices into enums

* Store castling related moves as consts

* SPRT result

```
Score of Simbelmyne vs simbelmyne-main: 458 - 377 - 596  [0.528] 1431
...      Simbelmyne playing White: 234 - 181 - 301  [0.537] 716
...      Simbelmyne playing Black: 224 - 196 - 295  [0.520] 715
...      White vs Black: 430 - 405 - 596  [0.509] 1431
Elo difference: 19.7 +/- 13.7, LOS: 99.7 %, DrawRatio: 41.6 %
SPRT: llr 2.99 (101.5%), lbound -2.94, ubound 2.94 - H1 was accepted
```

+++ chess/src/fen.rs
@@ -121,8 +121,8 @@ impl Board {
                     let bb = Bitboard::from(square);
 
                     piece_list[square_idx] = Some(piece);
-                    piece_bbs[piece.piece_type() as usize] |= bb;
-                    occupied_squares[piece.color() as usize] |= bb;
+                    piece_bbs[piece.piece_type()] |= bb;
+                    occupied_squares[piece.color()] |= bb;
 
                     square_idx += 1;
                 }
